pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"
[project]
name = "cvkit_optipose"
version = "0.0.1"
authors = [
    { name = "Mahir Patel", email = "mahirp@bu.edu" },
]
description = "Postural auto-encoder"
requires-python = ">=3.10"
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["bu_cvkit",
"matplotlib==3.7.1",
"numpy==1.24.2",
"scipy==1.10.1",
"tensorflow==2.11.0"]

[tool.setuptools]
packages = ["cvkit_optipose","OptiPose"]